# [SPARK-31443][SQL] Fix perf regression of toJavaDate