@cloud-fan 2020-04-15
The PR itself looks good but we should also investigate if 2.4 has a correctness issue.

Thanks, merging to master/3.0!