@MaxGekk 2020-04-15
> ... we should also investigate if 2.4 has a correctness issue

Here is the JIRA ticket for investigation: SPARK-31449